backend/requirements.txt
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
defusedxml==0.7.1
Django==3.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
pycparser==2.21
PyJWT==2.7.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.3